python/StegonagraphyPython/requirements.txt
PyQt5~=5.15.2
Pillow~=8.1.0